backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cloudinary==1.44.0
cryptography==46.0.6
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
google-auth==2.49.1
gunicorn==23.0.0
idna==3.11
packaging==26.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.5
stripe==15.0.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
